Koble til dna-api
